+++ tests/DfE.FindInformationAcademiesTrusts.UnitTests/Pages/OtherServicesLinkBuilderTests.cs
@@ -78,4 +78,16 @@ public class OtherServicesLinkBuilderTests
             _sut.FindSchoolPerformanceDataListingLink("1234", TrustType.SingleAcademyTrust, null);
         result.Should().BeNull();
     }
+
+    [Theory]
+    [InlineData("TR02345", "https://educationgovuk.sharepoint.com/_layouts/15/sharepoint.aspx?oobRefiners=%7B%22FileType%22%3A%5B%22other%22%5D%7D&q=TR02345&v=%2Fsearch")]
+    [InlineData("", "https://educationgovuk.sharepoint.com/_layouts/15/sharepoint.aspx?oobRefiners=%7B%22FileType%22%3A%5B%22other%22%5D%7D&q=&v=%2Fsearch")]
+    public void SharepointFolderLink_should_return_url_containing_GroupId(string trn, string expected)
+    {
+        var result =
+            _sut.SharepointFolderLink(trn);
+
+        result.Should()
+            .Be(expected);
+    }
 }